panion/plugins/world_model_manager/requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
networkx>=2.6.0
PyYAML>=6.0 